+++ src/main/java/onlyoffice/OnlyOfficeEditorServlet.java
@@ -204,6 +204,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
             editorConfigObject.put("lang", localeManager.getLocale(user).toLanguageTag());
             editorConfigObject.put("customization", customizationObject);
 
+            customizationObject.put("forcesave", configurationManager.forceSaveEnabled());
             customizationObject.put("goback", gobackObject);
             gobackObject.put("url", gobackUrl);