+++ src/main/java/onlyoffice/managers/configuration/ConfigurationManager.java
@@ -7,6 +7,7 @@ import java.util.Properties;
 public interface ConfigurationManager extends Serializable {
     public Properties getProperties() throws IOException;
     public String getProperty(String propertyName);
+    public boolean forceSaveEnabled();
     public boolean selectDemo(Boolean demo);
     public Boolean demoEnabled();
     public Boolean demoAvailable(Boolean forActivate);