+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log
 
+
 ## Added
 - open non-editable formats on views
 - added formats for conversion docm, dot, dotx, epub, htm, html, otp, ots, ott, pot, potm, potx, pps, ppsm, ppsx,
  pptm, rtf, xlsm, xlt, xltm, xltx
+- keep intermediate versions when editing (forcesave)
 - version history with highlighting changes
 
 ## 3.0.0